After working through the
problem for a night I created somewhat of a solution but it seems to fail on the last command. I implemented this is in a seperate file named pipelineTest where I explicitly created the commands array. I know that my program is correctly redirecting stdin and stdout in some cases but I wonder if the parent is correctly moving the output of one command into the input for the next. I have created a new file 'pipelineSegments' where I will isolate and test this functionality seperately. First, I will test if any data is being sent back through the pipe (from parent to child and vica versa).